% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R, R/card_unicode.R
\name{card_list}
\alias{card_list}
\alias{card_suit}
\alias{card_rank}
\alias{card_style}
\alias{card_color}
\alias{card_html}
\alias{card_specials}
\alias{card_htmlhex}
\alias{card_unicode}
\title{decks of 52 cards}
\usage{
card_list()

card_suit(card = card_list())

card_rank(card = card_list())

card_style(card = card_list(), four_color = TRUE)

card_color(card = card_list(), four_color = TRUE)

card_html(card = card_list(), size = 172, four_color = TRUE)

card_specials()

card_htmlhex(card = card_list())

card_unicode(card = card_list())
}
\arguments{
\item{card}{a card. A list of available cards can be found with \code{card_list()}}

\item{four_color}{should a four color deck be applied. This means that each
suit will have it's own color (black, red, blue and green). Otherwise, use
two colors (black and red).}

\item{size}{the font-size to use}
}
\description{
Functions that extract information from a card
}
\examples{
# show all valid cards
card_list()

# determine rank and suit and color
card_rank("Jc")
card_suit("7d")
card_color("Ad")
card_color("Ad", four_color = FALSE)

# style cards either in the console or via HTML
cat(card_unicode("5h"), "\\n")
cat(card_style("5h"), "\\n")
card_html("5h")

# iterate over several cards for styling
card_list() \%>\% sapply(card_style) \%>\% cat("\\n")
card_list() \%>\% lapply(card_html, 70) \%>\% span() \%>\% browsable()
}
