% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domino.R
\name{domino_unicode}
\alias{domino_unicode}
\alias{domino_html}
\title{domino pieces}
\usage{
domino_unicode(first = 0:6, second = 0:6, horizontal = TRUE)

domino_html(first = 0:6, second = 0:6, horizontal = TRUE,
  size = 172)
}
\arguments{
\item{first}{the first number to appear in the domino. Valid values: 0,1,...,6}

\item{second}{the second number to appear in the domino. Valid values: 0,1,...,6}

\item{horizontal}{whether to use horizontal or vertical alignment}

\item{size}{the font-size to use}
}
\description{
create Unicode representation of domino pieces and export to HTML.
}
\examples{
domino_html(5, 6)
domino_html(2, 0, horizontal = FALSE)
}
