% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draughts.R
\name{draughts_unicode}
\alias{draughts_unicode}
\alias{draughts_html}
\title{draughts pieces}
\usage{
draughts_unicode(color = c("white", "black"), status = c("man",
  "king"))

draughts_html(color = c("white", "black"), status = c("man", "king"),
  size = 172)
}
\arguments{
\item{color}{A color to be used. Black or white}

\item{status}{A status of the checker. Use \code{man} for a single
checker and `king`` for a double checker}

\item{size}{the font-size to use}
}
\description{
Functions to create Unicode draughts pieces and export them
to HTML
}
\examples{
span(draughts_html("white", "man"), draughts_html("white", "king"),
     draughts_html("black", "man"), draughts_html("black", "king")) \%>\% browsable()

}
