% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahjong.R
\name{mahjong_list}
\alias{mahjong_list}
\alias{mahjong_unicode}
\alias{mahjong_html}
\title{mahjong tiles}
\usage{
mahjong_list()

mahjong_unicode(tile = mahjong_list())

mahjong_html(tile = mahjong_list(), size = 172)
}
\arguments{
\item{tile}{a mahjong tile. Valid values can be found via \code{mahjong_list()}}

\item{size}{the font-size to use}
}
\description{
create unicode representations of mahjong tiles and export to HTML
}
\examples{
mahjong_html("autumn")
mahjong_html("bamboo", size = 400)

mahjong_list() \%>\% lapply(mahjong_html, 100) \%>\% span() \%>\% browsable()
}
