% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice_unicode}
\alias{dice_unicode}
\alias{dice_html}
\title{create six sided dice}
\usage{
dice_unicode(number = 1:6)

dice_html(number = 1:6, size = 172)
}
\arguments{
\item{number}{the number that should appear on the die}

\item{size}{the font-size to use}
}
\description{
create six sided dice
}
\examples{
dice_unicode(6)
dice_html(4)

1:6 \%>\% lapply(dice_html) \%>\% span() \%>\% browsable()
}
