% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{span}
\alias{div}
\alias{browsable}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmltools}{\code{\link[htmltools]{span}}, \code{\link[htmltools]{div}}, \code{\link[htmltools]{browsable}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

