% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit.R
\name{suit_list}
\alias{suit_list}
\alias{suit_unicode}
\alias{suit_color}
\alias{suit_style}
\alias{suit_html}
\title{Card suits}
\usage{
suit_list()

suit_unicode(suit = suit_list(), fill = TRUE)

suit_color(suit = suit_list(), four_color = TRUE)

suit_style(suit, four_color = TRUE)

suit_html(suit = suit_list(), size = 172, four_color = TRUE,
  fill = TRUE)
}
\arguments{
\item{suit}{one of the suits listed in \code{list_suits()}}

\item{fill}{whether to use a filled suit symbol like \code{BLACK SPADE SUIT} or an
unfilled one like \code{WHITE SPADE SUIT}}

\item{four_color}{should a four color deck be applied. This means that each
suit will have it's own color (black, red, blue and green). Otherwise, use
two colors (black and red).}

\item{size}{the font-size to use}
}
\description{
Functions that extract suit symbols and suit colors from a given suit
}
\examples{
# show all valid suits
suit_list()

# determine matching colors for the suit
suit_color("d")
suit_color("d", four_color = FALSE)

# show suits either in the console or as HTML
cat(suit_unicode("c"), "\\n")
suit_html("c")
suit_html("c", fill = FALSE)

# iterate over several suits for styling
suit_list() \%>\% sapply(suit_style) \%>\% cat("\\n")
suit_list() \%>\% lapply(suit_html) \%>\% span() \%>\% browsable()
suit_list() \%>\% lapply(suit_html, fill = FALSE) \%>\% span() \%>\% browsable()
}
