% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_unicode_html.R
\name{unicode_html}
\alias{unicode_html}
\title{convert a Unicode string into browsable html}
\usage{
unicode_html(x, size = 172, style = "")
}
\arguments{
\item{x}{a string to be wrapped}

\item{size}{the font-size to use}

\item{style}{a sting appended to the style attribute of the resulting \code{span}.}
}
\description{
convert a Unicode string into browsable html
}
